fixed behavior for goToJail()
the way it was before, the dice rolls weren't getting saved once you escaped jail, meaning it wouldn't move you the number of spaces you got while in jail.
how it works now is it checks to see you're out of jail and then does all the attempts to escape outside of the goToJail() function so that the rolls are saved. If you do get jailed, the doubleCounter is set to 0 which will trigger next() at the end of the function to move on to the next turn.